Switch Linux build from 32-bit to native 64-bit
- Remove -m32 flags from blakserv and blakcomp makefiles
- Fix DWORD/LPARAM typedefs to uint32_t (unsigned long is 8 bytes on 64-bit)
- Fix UINT4 in md5.h to uint32_t (fixes RSB hash mismatch with clients)
- Fix pointer-to-int casts in blakcomp with intptr_t (kodbase.c, codegen.c)
- Update install-build-deps.sh for native 64-bit packages

All binary formats (savegames, BOFs, client protocol) use explicit 4-byte
integers and remain fully compatible between 32-bit and 64-bit builds.
Tested with production savegame (5063 accounts, 0 errors).

+++ blakcomp/codegen.c
@@ -701,7 +701,7 @@ int codegen_switch(switch_stmt_type s, int numlocals)
 
    /* Backpatch continue statements in loop body */
    for (p = current_loop->for_continue_list; p != NULL; p = p->next)
-      BackpatchGotoUnconditional(outfile, (int)p->data, FileCurPos(outfile));
+      BackpatchGotoUnconditional(outfile, (intptr_t)p->data, FileCurPos(outfile));
 
    /* Go back and fill in destination address for unconditional goto */
    if (!default_stmt)
@@ -744,12 +744,12 @@ void codegen_exit_loop(void)
 
    /* Backpatch break statements to jump to end of loop */
    for (p = current_loop->break_list; p != NULL; p = p->next)
-      BackpatchGotoUnconditional(outfile, (int) p->data, FileCurPos(outfile));
+      BackpatchGotoUnconditional(outfile, (intptr_t) p->data, FileCurPos(outfile));
    current_loop->break_list = list_delete(current_loop->break_list);
 
    /* Backpatch conditional goto statements to jump to end of loop */
    for (p = current_loop->conditional_goto_list; p != NULL; p = p->next)
-      BackpatchGotoConditional(outfile, (int)p->data, FileCurPos(outfile));
+      BackpatchGotoConditional(outfile, (intptr_t)p->data, FileCurPos(outfile));
    current_loop->conditional_goto_list = list_delete(current_loop->conditional_goto_list);
 
    /* Remove current list from loop "stack" */
@@ -823,7 +823,7 @@ int codegen_dowhile(while_stmt_type s, int numlocals)
    
    /* Backpatch continue statements in loop body */
    for (p = current_loop->for_continue_list; p != NULL; p = p->next)
-      BackpatchGotoUnconditional(outfile, (int)p->data, FileCurPos(outfile));
+      BackpatchGotoUnconditional(outfile, (intptr_t)p->data, FileCurPos(outfile));
 
    numtemps = codegen_conditional_goto(s->condition, numlocals, &sourceval);
    if (numtemps > our_maxlocal)
@@ -896,7 +896,7 @@ int codegen_for(for_stmt_type s, int numlocals)
 
    /* Backpatch continue statements in loop body */
    for (p = current_loop->for_continue_list; p != NULL; p = p->next)
-      BackpatchGotoUnconditional(outfile, (int)p->data, FileCurPos(outfile));
+      BackpatchGotoUnconditional(outfile, (intptr_t)p->data, FileCurPos(outfile));
 
    /* Step 4: Execute statements from assign list (iterators) */
    /* If no iteration, list will be NULL. */
@@ -990,7 +990,7 @@ int codegen_foreach(foreach_stmt_type s, int numlocals)
 
    /* Backpatch continue statements in loop body */
    for (p = current_loop->for_continue_list; p != NULL; p = p->next)
-      BackpatchGotoUnconditional(outfile, (int)p->data, FileCurPos(outfile));
+      BackpatchGotoUnconditional(outfile, (intptr_t)p->data, FileCurPos(outfile));
 
    /**** Statement #4:    temp = Rest(temp) ****/
    /* Can reuse temp_expr from statement #3 above */
